rad-assistant/requirements.txt
pydantic>=2.0
requests>=2.28
python-dotenv>=0.20
fastapi>=0.95
uvicorn[standard]>=0.22
httpx>=0.24
torch
monai
pydicom
opencv-python
numpy
Pillow
pytest
python-multipart
torchxrayvision
